+++ ceammc/ext/src/ui/ui_toggle.h
@@ -7,6 +7,8 @@ using namespace ceammc;
 
 class UIToggle : public UIObject {
     t_rgba prop_color_active;
+    t_float prop_value_on_;
+    t_float prop_value_off_;
     bool value_;
 
 public: